requirements.txt
streamlit
python-dotenv
PyYAML
google-generativeai
sentence-transformers
faiss-cpu
numpy
scikit-learn